+++ nsupdate/settings/dev.py
@@ -7,9 +7,9 @@ from .base import *
 DEBUG = True
 TEMPLATE_DEBUG = DEBUG
 
-WE_HAVE_SSL = False  # True if you run a https site also, suggest that site to users if they work on the http site.
-CSRF_COOKIE_SECURE = WE_HAVE_SSL
-SESSION_COOKIE_SECURE = WE_HAVE_SSL
+WE_HAVE_TLS = False  # True if you run a https site also, suggest that site to users if they work on the http site.
+CSRF_COOKIE_SECURE = WE_HAVE_TLS
+SESSION_COOKIE_SECURE = WE_HAVE_TLS
 
 BASEDOMAIN = 'nsupdate.info'
 WWW_HOST = 'localhost:8000'